Refactor predictions
