+++ src/anomalib/models/ensemble/predictions/prediction_data.py
